Deferred properties: avoid spurious binding loop
During binding evaluation, it might happen that we do an initial read of
a deferred property. If that deferred property is in turn an object
property, further binding evaluations might occur. Those bindings should
not depend on whatever trigerred the initial binding evaluation.
However, with the new C++ properties/bindings, that would happen, as the
bindingStatus would still indicate that binding evaluation is active.

To remedy this, we temporarily suspend the binding status ind begin and
completeDeferred.

Pick-to: 6.5 6.4 6.2
Fixes: QTBUG-96351
Change-Id: I7e9778aa5339f044ce3082e962a2ede34b5d1e69
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>