Ignore DPR for image providers if the requested size is not set
If the requested size is not set, the QQuickImageBasePrivate::sourcesize
is -1. Multiplying it with devicePixelRatio is useless because it will
be still an invalid size and it will be ignored. Nonetheless,
QQuickImageBase::pixmapChange() will still divide the size with
devicePixelRatio resulting smaller images.

The issue is reproducible with the tst_qquickimageprovider auto test if
ran with QT_SCALE_FACTOR=2 set.

Pick-to: 6.2 6.4 6.5
Fixes: QTBUG-109854
Change-Id: I9fea39c34144e5eae7e421bccc8dc65388dd5b32
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>