Fix connection leak in header proxy model
In header view, the header proxy model which acts as proxy for table
model, connects headerDataChanged() signal to a slot with both sender
and receiver as table model object. This connection happens each time
header view instantiates or model in the header view need to be reset.

In a case, where header view reinstantiated (released and loaded) and
table model not released, the newly instantiated header proxy model
creates a connection to the existing table model for headerDataChanged()
signal and this creates a connection leak.

This can be avoided by making header proxy model as receiver
similar to other slots within the proxy model.

Fixes: QTBUG-108610
Pick-to: 6.2 6.4 6.5
Change-Id: I1b250f98310e16743f7fc2c4cc11cbfb3d9fff45
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>