HeaderView: ensure that the implicit size is greater than zero
HorizontalHeaderView should have the same implicit height as the
delegate items it contains. And this height is reflected by
contentHeight. The problem is that the content height is zero at
start-up, before the view gets populated with delegate items.
The result is that the height of the view will then be zero at
start-up, which can lead to TableView not loading any
items at all.

This patch will therefore ensure that the implicit height of
HeaderView is 1 at start-up, so that the delegate items will be
loaded. Once loaded, the implicit height of the view will be
adjusted to the correct value.

Pick-to: 6.5
Change-Id: I5181840665f648275e28dd8fe943e759797388fa
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>