Make quick examples that rely on the shared files build with qmake again
In d270c51f812b5f0ffe6208b38c8ea0e898535e50 the DECLARATIVE_EXAMPLE_MAIN
macro was modified to prepend the "qrc:/qt/qml" prefix to file paths that
should be loaded by the engine. But the patch forgot to update the
shared.qrc prefix, which means that the examples that relied on the
shared module, would no longer build out-of-the-box with qmake.

This patch solve the issue, so that the examples work with qmake
once again.

Pick-to: 6.5
Change-Id: I677c1e61423cf98c6a1030ce591707888456ef19
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>