Fix Shape stroke color when changing alpha from 0 to non-0
The handling of the stroke color lacks the extra checks that are
performed when setting the fill color. Switching from "transparent"
to another color involves setting not just the color-dirty flag but
also the geometry-dirty since alpha == 0 may skip operations such as
triangulating, meaning the work has to be performed later when once
again switching to an alpha > 0 color.

Change-Id: Ibd6d762cfcb07ca8b9c852553c979bcb4ee9fa21
Fixes: QTBUG-109882
Pick-to: 6.5 6.4 6.2 5.15
Reviewed-by: Christian Strømme <christian.stromme@qt.io>