tst_QQuickListView2::flickDuringFlicking: allow flicking less
The dynamic behavior of Flickable's animation will not stay fixed:
we want to make deceleration platform-dependent; and it turns out that
this test was expecting it to flick quite far after a series of touch
events, whereas if it decelerates faster, it does not get quite so far.
The exact amount doesn't really matter.

Pick-to: 6.5
Change-Id: I6b3507faf21aaaea72c5bed6b2b89d5530264310
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>