QQuickPopup: fix logic for modally blocking wheel events
After caca7d7d4fcb7281332904c01950a66385bad0b3, QQuickOverlay ate wheel
events to any item that was not a child of the top-most modal popup.
This broke event delivery to children of modless popups higher up in the
stack.

Fix that logic so that we don't eat events if we find a popup that the
item is a child of. Only eat the event if we don't find such a popup
before hitting the first modal popup.

Add a test with a relevant scenario, and as a drive-by, fix and simplify
the helper function sending wheel events. Since the event gets delivered
to the window, we have to use scene coordinates, not item coordinates.
This happened to work in the wheel test case as the test sliders cover
the entire window.

Fixes: QTBUG-110023
Pick-to: 6.5 6.4
Change-Id: Ibc53114cf09693e23b1e66f28a0c2056f3bd810e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>