iOS: add horizontal and vertical headerview controls
The current implementation would pick up the headerview
controls from the Basic style (fallback). The problem is
that those controls are hard-coded to always be in Light
mode.

This patch will therefore add HorizontalHeaderView.qml
and VerticalHeaderView.qml to the iOS style, and use the
palette to ensure that the headers looks correct, also
in Dark mode.

Pick-to: 6.5
Change-Id: Ic91234bae09bc73b9c1cbd2fc439d2faa23a9b72
Reviewed-by: Doris Verria <doris.verria@qt.io>