qmlformat: fix omitting some comments while reformatting
We rewrite comments associated to a node on the preVisit call
(if they were marked as preComment), or postVisit( if comments were
marked as postComments) of the reformatter. If the comment
associated with a patternProperty kind of node, neither of these
functions are called. Add missing call to previsit/postVist
in the pattern property node visit.

Pick-to: 6.4 6.5
Fixes: QTBUG-109074
Change-Id: If57968b3f5dbd83aa23dc2cd2bca3608ee841d49
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>