Have a way to always use an elapsed time based animation driver
Set QSG_USE_SIMPLE_ANIMATION_DRIVER=1 to get an alternative animation
driver. (applies when using the threaded render loop) This behaves like
the fallback mode (TimerMode) of the other one: it advances based on
whatever delta the QElapsedTimer reports. Thus there is no dependency
on vsync-based throttling whatsoever, and a big chunk of the workaround
infrastructure (such as needing to fall back to QTimer with multiple
windows) is simply not needed.

It should be evaluated going forward if this should become the default.
For now, in order to avoid any surprises, it is only opt-in.

Task-number: QTBUG-84328
Change-Id: I8e466caa08ba5f0471d767b1600d198ba787d8e5
Reviewed-by: Janne Koskinen <janne.p.koskinen@qt.io>
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit ac5bbbe4aca5e1abef4972aeeba22c8c2f923705)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>