tst_QmllsCompletions: rename into tst_qmlls_modules
Most of tst_QmllsCompletions logic to start and feed data to an instance
of qmlls can be reused to test other modules, so rename the test so it
can be reused for, e.g., the go-to type definition module.

This commits only renames the test into tst_qmlls_modules and the tests
folders and moves the test's declaration in a header file. Also
renames the target in the CMakeLists as tst_qmlls_modules.

Task-number: QTBUG-109006
Change-Id: I4c20575a89a6193016963de3a87c4c6ec08e4559
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>