Bring QSurfaceFormat configurability to qmlscene's level in the qml tool
Restore the --transparent and --multisample options.  Also make the
default QSurfaceFormat setting universal (no need to tie it to the
option of forcing a core profile context), and prevent the warning
that is printed when the default surfaceformat is changed after the
application is already created with AA_ShareOpenGLContexts set. This
involves having to handle the relevant arguments manually early on,
not through QCommandLineParser.

All this matches the qmlscene behavior.

NB qmlscene requests 16 samples with --multisample which is completely
pointless in practice. Use the common 4 now.

Change-Id: I671e5f22846a715675ea936bb7cf2a31a849dc28
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>