tst_dialog: Fix testing for implicitWidth and wait for enter transitions
- The Dialog’s implicitWidth is set to the max between the header’s,
footer’s, and background’s implicitWidth. So, when testing that the
dialog’s implicitWidth is equal to that of the header, we should first
make sure that the IW of the background and footer is smaller.

- Wait for all enter transitions to finish before assuming that the
dialog is opened.

Change-Id: I803db9664ad46a83dbe6b47c53448530c30d6150
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 4a9be82a382b58155949ad4ec75e8ce4e12c47e6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>