MultiPointTouchArea: rename signal arguments in Qt 7; test with formals
This will be a behavior change for anyone who is still relying on
"touchPoints" to be the signal argument rather than the MPTA property;
so let's put it off until Qt 7 and have a reminder in docs in the
meantime.

Fix autotests to avoid depending on this (just as users should).

[ChangeLog][QtQuick][MultiPointTouchArea] As a reminder: you should
always use formal parameters when writing a signal handler that depends
on an argument, such as MultiPointTouchArea.onPressed, onUpdated,
onReleased, or onCanceled. It gives you an opportunity to name the
argument, to avoid confusion with the touchPoints property, and will
avoid behavior changes if we rename the argument in a future version.

Fixes: QTBUG-110906
Change-Id: Ie78dc539f7f6cac660f2e90bf3e4665afed3e3f7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 2b6eb1a51d1593b3df55b5ab4aee3d0eadd75bef)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>