qmlls_utils::findTypeIn and baseObject
Add two static convenience methods to the qmlls_utils.
baseObject() resolves the base object of an QmlObject, while
findTypeDefinitionOf() allows to extract the type of bindings, ids,
property definitions and method parameters.

Write tests for it.

Also, fix the MethodParameters to visit the correct typepath, instead of
the CppTypePath. This allows to find MethodInfos when asking for type
definitions of parameter and return types in functions (but just in the
"function f(x: MyComponent): MyComponent{" part, not tested on the
scriptexpression inside the function).

Task-number: QTBUG-110309
Change-Id: I6307f934fb7740891a13fd0a8f68466f1daaac87
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>