Rewrite qmlls_utils::itemsFromTextLocation for properties
It seems that itemsFromTextLocation is not working as expected.
The old version was taking the item that is nested at the biggest depth.
It turns out this fails for inline components, as main components and
inline components are placed on the same depth on the DomItem
construction.

( Amends c3e1ee45c30c6380e5c1d5b542b2951ea692bb52. )

Also, this does not work in presence of property maps: they stretch from
the first property to the last but have a nested structure that gives
them a bigger depth than the actual items it encloses. This means items
lying between two property definitions cannot be found with
itemsFromTextLocation, as it only returns the property map.

Rewrite qmlls_utils::itemsFromTextLocation to ignore the depth, and
instead always choose the item with the smallest coderegion,
independently of their depth. This solves the problem of main
component shadowing their inline components and property maps
shadowing everything between their properties.

Move the tst_qmlls_utils::createEnvironmentAndLoadFile implementations to
the top of the tst_qmlls_utils.cpp file, instead of being between
findItemFromLocation_data() and findItemFromLocation().

Task-number: QTBUG-110309
Change-Id: I3485b222de6a623cdff2a2b33035ab1dc96d79fb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>