DialogButtonBox: Add note about the alignment property in the docs
The DialogButtonBox implementation always assumes that the buttons are
laid out horizontally. The layout update based on the button alignment
is also done considering a horizontal orientation. Add a note on this
in the docs for the iOS style, which uses a vertical orientation too.

As a proper fix, we might want to consider adding an orientation
property to the DialogButtonBox for this case.

Change-Id: I36a80f1bb65a0c41da09816c2bf516c1c6943a4b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 89e297662ed86298485c935c64ef4d6b60edc260)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>