Unexport QSGMaterialShader::GraphicsPipelineState
[ChangeLog] The GraphicsPipelineState struct in QSGMaterialShader is no
longer exported. This may present a binary compatibility break in
theory, but unlikely to affect any applications in practice given
that there is no use case for creating or copying an instance of
this type in user code. This struct serves solely as input to the
virtual updateGraphicsPipelineState() function, and instances are
always created and initialized with default values by Qt.

Pick-to: 6.5
Change-Id: I5727ce6bb6ae4a86a44124100550e79b399a4ccf
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>