QML: Don't unconditionally invalidate cache if inline component is used
With inline components, we might have the case that an outer components
depends on an inline component defined in the same file. In that case,
the ResolvedType of the inline component has some very specific
characteristics we can recognize when adding to the hash.

We don't actually have to add  such inline components to the hash since
they are just part of the same file. If they change, the file will
change, leading to a different timestamp, which is caught earlier.

Pick-to: 6.5 6.5.0 6.2
Fixes: QTBUG-111042
Change-Id: I8ae716e55dd783cc8409039bd7baffe051df2b96
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>