JIT: Fix isNullOrUndefined bad logic
A logic error was left over in isNullOrUndefined that didn't actually
check for null of undefined after a change of encoding.

tst_qqmllanguage::isNullOrUndefined() was added to check the bug doesn't
happen anymore. Two cases are tested. The first is intended to run on
the interpreter. It was fine before this change and should be fine
afterwards. The second one calls the same function many times for JIT
optimizations to kick in and compile the function. This should ensure
the broken '??' and '?.' operators are fixed when running on the JIT.

Amends c7722d4ed61d6a887e9f6c403ffa10b2048de2a4.

Fixes: QTBUG-111088
Change-Id: I0e6d77363770e801aa586588e242220dec9d5e0a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit e5c2b236a04050bf20aaf96068e9b9b911bc70ad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>