QQuickSwipe: Fix memory leak
SwipeDelegate.swipe instantiates various Components. It does however not
track their lifetime, which so far led to memory leaks.
We could use unique_ptr to tie the lifetime of the items to swipe,
however that might break existing code if it were to do
property Item someItem: swDelegate.swipe.leftItem
// delete swDelegate

While unlikely, we can avoid that potential problem by simply letting
the engine manage the lifetime of the delegate (as we set control as the
parentItem, the garbage collector will not collect it while control
still exists, compare QQuickItemWrapper::markObjects.

Pick-to: 6.5
Change-Id: Ib1f87b053b23e46a1346cdd7b4c926fc79b08883
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>