QQuickStackView:: Avoid memory leak in QQuickStackView::push
If duplicate elements are pushed, they immediately get removed. However,
the list contains heap-allocated QQuickStackElements, so we actually
need to free them.
Fix that by porting the (accidentally quadratic) loop with a combination
of remove_if + delete + erase.

A cleaner but more invasive solution would have been to port to a vector
of unique_ptrs (which would have enabled usage of  erase_if), but that
is left for another patch.

Pick-to: 6.5
Change-Id: I0545cc4f7ccbf70b58e09cd8f8f06216d984eae3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>