Add QML API for application permissions
The API is based on dedicated QML types for each permission type,
where each instance of the permission reports status separately.

Requests for permissions are done imperatively for now, as the
semantics of a declarative request were hard to nail down, and
the user can easily build a declarative flow on top of the
imperative API that fits their needs and semantics.

Task-number: QTBUG-111001
Done-with: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Change-Id: I0ebc99e0fa0d22b0af5aa1d34851c5f0c62357c0
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>